rename base.js
