+++ lib/base.js
@@ -1,84 +1,47 @@
+const Error = require('./Error')
+const constants = require('./constants')
 /**
- * @typedef {Object} WeebImageReponse
- * @property {Number} status - The returned status code for the request made
- * @property {String} id - The internal ID of this image
- * @property {String} type - The type this image belongs to
- * @property {String} baseType - The bas Type this image belongs to
- * @property {Boolean} nsfw - Whenever or not this image is not suitable for work
- * @property {String} fileType - The file extension of this image
- * @property {String} mimeType - The mime typing ob this image
- * @property {string[]} tags - Array of tags that this image has
- * @property {String} url - The publicly reachable URL for this image
- * @property {Boolean} hidden - Whenever or not this image is hidden
- * @property {String} account - Account ID of the user that uploaded this image
+ * Main Handler for Weeb.sh
  */
+class Base {
+  /**
+   * @class Base - Main handler for all Weeb.sh request
+   */
+  constructor () {
+    /**
+     * Error constructor
+     * @type {WeebJSError}
+     */
+    this.Error = Error
+  }
 
-/**
- * @typedef {Object} WeebTypesResponse
- * @property {Number} status - The returned status code for the request made
- * @property {string[]} types - Array of all available types
- * @property {WeebImagePreview[]} [preview] - A preview object for every type
- */
-
-/**
- * @typedef {Object} WeebImagePreview
- * @property {String} id - The internal ID of this image
- * @property {String} type - The type this image belongs to
- * @property {String} baseType - The bas Type this image belongs to
- * @property {String} fileType - The file extension of this image
- * @property {String} url - The publicly reachable URL for this image
- */
-
-/**
- *  @typedef {Object} WeebTagsResponse
- * @property {Number} status - The returned status code for the request made
- * @property {string[]} tags - Array of all available tags
- */
-
-/**
- * Base URL for all API requests
- * @default
- * @constant
- * @type {string}
- */
-const BASE_URL = 'https://api.weeb.sh'
-
-/**
- * All available Endpoints to be used
- * @default
- * @type {Object}
- */
-const ENDPOINTS = {
-  tags: '/images/tags',
-  types: '/images/types',
-  random: '/images/random',
-  generate: '/auto-image/generate',
-  discord_status: '/auto-image/discord-status',
-  license: '/auto-image/license',
-  waifu_insult: '/auto-image/waifu-insult',
-  love_ship: '/auto-image/love-ship'
+  /**
+   * Send a request to Weeb.sh
+   * @param {String} endpoint - The endpoint to make the request to
+   * @param {Object} query - Query string params or Post options if method is post
+   * @param {Boolean} [isPost=false] - Whenever this is a post or a get
+   * @return {Promise<Buffer|Object>}
+   */
+  requestAPI (endpoint, query = {}, isPost = false) {
+    return new Promise(async (resolve, reject) => {
+      try {
+        let res
+        if (isPost) {
+          res = await this.axios.post(constants.ENDPOINTS[endpoint], query, {
+            headers: {'Content-Type': 'application/json'},
+            responseType: 'arraybuffer'
+          })
+        } else {
+          res = await this.axios.get(constants.ENDPOINTS[endpoint], {params: query, responseType: ['discord_status', 'generate'].includes(endpoint) ? 'arraybuffer' : null})
+        }
+        let {data, status} = res
+        if (status < 200 || status > 300) reject(Buffer.isBuffer(data) ? data.toString() : data)
+        else resolve(data)
+      } catch (err) {
+        reject(err)
+      }
+    })
+  }
 }
-/**
- * All available file types to be used in {@link WeebHandler#getRandom}
- * @default
- * @type {string[]}
- */
-const FILE_TYPES = [
-  'jpeg',
-  'jpg',
-  'png',
-  'gif'
-]
-/**
- * All available discord status types to be used in {@link WeebHandler#generateDiscordMock}
- * @default
- * @type {string[]}
- */
-const DISCORD_STATUSES = [
-  'online',
-  'idle',
-  'dnd',
-  'offline',
-  'streaming'
-]
-module.exports = {BASE_URL, ENDPOINTS, FILE_TYPES, DISCORD_STATUSES}
+
+module.exports = Base
